Analyse Verbessert

+++ src/main/java/de/mankianer/gruppe5/analysis/AnalyseKeySelector.java
@@ -3,10 +3,10 @@ package de.mankianer.gruppe5.analysis;
 import de.mankianer.gruppe5.model.analyse.Analyse;
 import org.apache.flink.api.java.functions.KeySelector;
 
-public class AnalyseKeySelector implements KeySelector<Analyse, Integer> {
+public class AnalyseKeySelector implements KeySelector<Analyse, Long> {
 
   @Override
-  public Integer getKey(Analyse analyse) throws Exception {
+  public Long getKey(Analyse analyse) throws Exception {
     return analyse.getTweetID();
   }
 }
